Added Movie Service Backend Test + Fixed bugs uncovered by Tests
